+++ src/ProcessMaker/Nayra/Bpmn/Models/DataStore.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProcessMaker\Nayra\Bpmn\Model;
+namespace ProcessMaker\Nayra\Bpmn\Models;
 
 use ProcessMaker\Nayra\Bpmn\DataStoreTrait;
 use ProcessMaker\Nayra\Contracts\Bpmn\DataStoreInterface;
