+++ src/ProcessMaker/Nayra/Bpmn/Models/EventCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProcessMaker\Nayra\Bpmn\Model;
+namespace ProcessMaker\Nayra\Bpmn\Models;
 
 use ProcessMaker\Nayra\Bpmn\Collection;
 use ProcessMaker\Nayra\Contracts\Bpmn\EventCollectionInterface;
