+++ src/ProcessMaker/Nayra/Bpmn/Models/FlowCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProcessMaker\Nayra\Bpmn\Model;
+namespace ProcessMaker\Nayra\Bpmn\Models;
 
 
 use ProcessMaker\Nayra\Bpmn\Collection;
