+++ src/ProcessMaker/Nayra/Bpmn/Models/StartEvent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProcessMaker\Nayra\Bpmn\Model;
+namespace ProcessMaker\Nayra\Bpmn\Models;
 
 use ProcessMaker\Nayra\Bpmn\StartEventTrait;
 use ProcessMaker\Nayra\Contracts\Bpmn\StartEventInterface;
