+++ src/ProcessMaker/Nayra/Bpmn/Models/Token.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ProcessMaker\Nayra\Bpmn\Model;
+namespace ProcessMaker\Nayra\Bpmn\Models;
 
 
 use ProcessMaker\Nayra\Bpmn\TokenTrait;
