+++ tests/Feature/Engine/ActivityExceptionTest.php
@@ -3,7 +3,7 @@
 namespace Tests\Feature\Engine;
 
 use ProcessMaker\Nayra\Bpmn\ActivityTrait;
-use ProcessMaker\Nayra\Bpmn\Model\ActivityActivatedEvent;
+use ProcessMaker\Nayra\Bpmn\Models\ActivityActivatedEvent;
 use ProcessMaker\Nayra\Contracts\Bpmn\ActivityInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\DataStoreInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\EndEventInterface;
