+++ tests/Feature/Engine/MessageStartEventTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature\Engine;
 
-use ProcessMaker\Nayra\Bpmn\Model\DataStoreCollection;
+use ProcessMaker\Nayra\Bpmn\Models\DataStoreCollection;
 use ProcessMaker\Nayra\Bpmn\Models\Collaboration;
 use ProcessMaker\Nayra\Bpmn\Models\Participant;
 use ProcessMaker\Nayra\Contracts\Bpmn\ActivityInterface;
