+++ tests/ProcessMaker/Models/CallActivity.php
@@ -3,7 +3,7 @@
 namespace ProcessMaker\Models;
 
 use ProcessMaker\Nayra\Bpmn\ActivitySubProcessTrait;
-use ProcessMaker\Nayra\Bpmn\Model\ActivityActivatedEvent;
+use ProcessMaker\Nayra\Bpmn\Models\ActivityActivatedEvent;
 use ProcessMaker\Nayra\Contracts\Bpmn\ActivityInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\CallActivityInterface;
 use ProcessMaker\Nayra\Contracts\Bpmn\CallableElementInterface;
