+++ tests/unit/ProcessMaker/Nayra/Bpmn/BaseTraitTest.php
@@ -3,7 +3,7 @@
 namespace ProcessMaker\Nayra\Bpmn;
 
 use PHPUnit\Framework\TestCase;
-use ProcessMaker\Nayra\Bpmn\Model\Process;
+use ProcessMaker\Nayra\Bpmn\Models\Process;
 use ProcessMaker\Nayra\Contracts\Bpmn\EntityInterface;
 
 class BaseTraitTest extends TestCase
